Create Collatz Sequences
Broadly, number theory concerns itself with the properties of the positive integers. G.H.
Hardy was a brilliant British mathematician who lived through both World Wars and
conducted a large deal of number-theoretic research. He was also a pacifist who was
happy that, from his perspective, his research was not “useful”. He wrote in his 1940
essay A Mathematician’s Apology “[n]o one has yet discovered any warlike purpose to be
served by the theory of numbers or relativity, and it seems very unlikely that anyone will
do so for many years.”1 Little did he know, the purest mathematical ideas of number
theory would soon become indispensable for the cryptographic techniques that kept
communications secure. Our subject here is not number theory, but we will see a few
times where combinatorial techniques are of use in number theory.
Example 1.4. Form a sequence of positive integers using the following rules. Start with
a positive integer n > 1. If n is odd, then the next number is 3n + 1. If n is even, then
the next number is n/2. Halt if you ever reach 1. For example, if we start with 28, the
sequence is
28, 14, 7, 22, 11, 34, 17, 52, 26, 13, 40, 20, 10, 5, 16, 8, 4, 2, 1.

Now suppose you start with 19. Then the first few terms are
19, 58, 29, 88, 44, 22.
But now we note that the integer 22 appears in the first sequence, so the two sequences
will agree from this point on. Sequences formed by this rule are called Collatz sequences.
Pick a number somewhere between 100 and 200 and write down the sequence you
get. Regardless of your choice, you will eventually halt with a 1.